api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg[binary]
pydantic
pydantic-settings
email-validator
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
python-json-logger
pytest
httpx